split some functions
