+++ Makefile
@@ -10,9 +10,11 @@ SRC_DIR = src
 OBJS_DIR = objs
 
 SRC =  parsing.c \
-		execute.c \
 		main.c \
 		signals.c \
+		execution/execute.c \
+		execution/build_args.c \
+		execution/find_cmd.c \
 		envp/envp_errors.c \
 		envp/envp_to_str.c \
 		envp/envp_utils.c \
