+++ src/execution/execute.c
@@ -6,55 +6,11 @@
 /*   By: abdsalah <abdsalah@student.42amman.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/30 01:23:07 by abdsalah          #+#    #+#             */
-/*   Updated: 2025/02/06 02:52:16 by abdsalah         ###   ########.fr       */
+/*   Updated: 2025/02/06 06:23:38 by abdsalah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
-#include <stdio.h>
-#include <unistd.h>
-
-static int is_redirect_operator(int type) {
-    return (type == REDIRECT_IN ||
-            type == REDIRECT_OUT ||
-            type == REDIRECT_APPEND ||
-            type == HEREDOC);
-}
-
-static char **build_command_argv(t_command *cmd) 
-{
-    int i = 0, j = 0, count = 0;
-    while (i < cmd->token_count)
-	{
-        if (is_redirect_operator(cmd->tokens[i]->type))
-            i += 2; 
-        else {
-            count++;
-            i++;
-        }
-    }
-    char **argv = malloc(sizeof(char *) * (count + 1));
-    if (!argv) return NULL;
-    i = 0;
-    j = 0;
-    while (i < cmd->token_count)
-	{
-        if (is_redirect_operator(cmd->tokens[i]->type)) {
-            i += 2;
-            continue;
-        }
-        argv[j] = ft_strdup(cmd->tokens[i]->value);
-        if (!argv[j]) {
-            free_str_array(argv);
-            return NULL;
-        }
-        j++;
-        i++;
-    }
-    argv[j] = NULL;
-    return argv;
-}
-
 
 static int	get_redirections(t_command *cmd, int *in_fd, int *out_fd)
 {
@@ -175,94 +131,6 @@ void	free_str_array(char **arr)
 	free(arr);
 }
 
-static char *find_executable_in_paths(char **cmd, char **paths, char **full_path, int *i)
-{
-    while (paths[*i])
-    {
-        *full_path = malloc(ft_strlen(paths[*i]) + ft_strlen(*cmd) + 2);
-        if (!*full_path)
-            return (NULL);
-        sprintf(*full_path, "%s/%s", paths[*i], *cmd);
-        if (access(*full_path, X_OK) == 0)
-        {
-            free_str_array(paths);
-            return (*full_path);
-        }
-        free(*full_path);
-        (*i)++;
-    }
-    return (NULL);
-}
-
-static char *find_command_path(char *cmd, char **envp)
-{
-    char    *path_var;
-    char    **paths;
-    char    *full_path;
-    int     i;
-
-    if (ft_strchr(cmd, '/'))
-        return (ft_strdup(cmd));
-    i = -1;
-    while (envp[++i])
-    {
-        if (ft_strncmp(envp[i], "PATH=", 5) == 0)
-        {
-            path_var = envp[i] + 5;
-            paths = ft_split(path_var, ':');
-            if (!paths)
-                return (NULL);
-            i = 0;
-            full_path = find_executable_in_paths(&cmd, paths, &full_path, &i);
-            if (full_path)
-                return (full_path);
-            free_str_array(paths);
-            return (NULL);
-        }
-    }
-    return (NULL);
-}
-int builtins(char **arg)
-{
-	if (ft_strcmp(arg[0], "echo") == 0)
-		return (1);
-	if (ft_strcmp(arg[0], "cd") == 0)
-		return (1);
-	if (ft_strcmp(arg[0], "pwd") == 0)
-		return (1);
-	if (ft_strcmp(arg[0], "export") == 0)
-		return (1);
-	if (ft_strcmp(arg[0], "unset") == 0)
-		return (1);
-	if (ft_strcmp(arg[0], "env") == 0)
-		return (1);
-	if (ft_strcmp(arg[0], "exit") == 0)
-		return (1);
-	return (0);
-
-}
-
-
-
-int exec_builtins(char **args, t_shell *shell)
-{
-	if (ft_strcmp(args[0], "echo") == 0)
-		return (ft_echo(args));
-	if (ft_strcmp(args[0], "cd") == 0)
-		return (ft_cd(args + 1, &shell->env_list));
-	if (ft_strcmp(args[0], "pwd") == 0)
-		return (ft_pwd());
-	if (ft_strcmp(args[0], "export") == 0)
-		return (ft_export(args + 1, &shell->env_list));
-	if (ft_strcmp(args[0], "unset") == 0)
-		return (ft_unset(args + 1, &shell->env_list));
-	if (ft_strcmp(args[0], "env") == 0)
-		return (ft_env(args + 1, shell));
-	if (ft_strcmp(args[0], "exit") == 0)
-		return (ft_exit(args + 1, shell));
-	return (0);
-
-}
 
 void execute_pipeline(t_shell **shell)
 {
@@ -302,7 +170,7 @@ void execute_pipeline(t_shell **shell)
             fprintf(stderr, "minishell: invalid command\n");
             exit(EXIT_FAILURE);
         }
-        if (builtins(argv) && (*shell)->parser->command_count == 1)
+        if (is_builtin(argv) && (*shell)->parser->command_count == 1)
         {
             int fd1 = -1;
             int fd2 = -1;
@@ -363,7 +231,7 @@ void execute_pipeline(t_shell **shell)
                     fprintf(stderr, "minishell: invalid command\n");
                     exit(EXIT_FAILURE);
                 }
-                if (builtins(argv))
+                if (is_builtin(argv))
                 {
                     (*shell)->exit_status = exec_builtins(argv, *shell);
                     free_str_array(argv);
